signature: change the option used to specify the secret key
Before the patch to specify the secrete key option "--default-key" was used.
But it's not so clear for 2 reasons:
- If the specified key is absent, the current default key will be used.
- We don't want to change the default secrete key, we want to sign
  some data by specified key.
Now we are using the "-u" option. As a result: if the specified key
is absent, an error will be returned and the key used by default is
not change now.

Part of #19