options: add force mode
Now, when adding packages to the index, if an malformed package
is present, the program will be terminated. In some cases, the
desired program behavior is to correctly create an index with all
correct packages and display a message about the presence of
malformed ones.
This behavior can now be set using the "--force" option.
The malformed_list.txt file will also be added to the repository.
For rpm based:
`repodata/malformed_list.txt`
For deb based:
`dists/dist_name/malformed_list.txt`