rpm: fix architecture definition for srpm
Before the patch, the architecture of the package was written in
metadata for srpm, but this is not correct. For srpm, the
archetecture field must be filled with "src".

Tip:
To view the value of some tag in the rpm header use:
`rpm --query --package package.src.rpm --queryformat "%{TAG_NAME}\n"`

Related to tarantool/tarantool-qa#148