rpm: support deletion of stale repo metadata
The repo metadata needs to be updated when RPM packages are deleted from
s3 backend for a some reason. Now `mkrepo` supports not only adding, but
also deleting the metadata and does it in the same iteration. To make it
possible, `mkrepo` deletes the metadata first, then adds it.

Part of #50